Fixed some bugs

+++ Api/ActionLog/LogTemporaryOpeningLockedPayload.cs
@@ -5,7 +5,7 @@ namespace ChasterSharp
     public sealed class LogTemporaryOpeningLockedPayload
     {
         [JsonPropertyName("unlockedTime")]
-        public int UnlockedTime { get; set; }
+        public float UnlockedTime { get; set; }
 
         [JsonPropertyName("penaltyTime")]
         public int PenaltyTime { get; set; }
